refactor: resharper code cleanup
