+++ src/Tool/Git2SemVer.Tool.Tests/Commands/Remove/RemoveCommandTests.cs
@@ -3,8 +3,7 @@ using NoeticTools.Git2SemVer.Core.Console;
 using NoeticTools.Git2SemVer.Core.Tools;
 using NoeticTools.Git2SemVer.Core.Tools.DotnetCli;
 using NoeticTools.Git2SemVer.Testing.Core;
-using NoeticTools.Git2SemVer.Tool.Commands.Changelog;
-using NoeticTools.Git2SemVer.Tool.Commands.Remove;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Remove;
 using NoeticTools.Git2SemVer.Tool.MSBuild.Solutions;
