+++ src/Tool/Git2SemVer.Tool/CommandLine/Git2SemVerCommandApp.cs
@@ -1,6 +1,6 @@
 ﻿using NoeticTools.Git2SemVer.Core.Logging;
-using NoeticTools.Git2SemVer.Tool.Commands.Add;
-using NoeticTools.Git2SemVer.Tool.Commands.Remove;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Add;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Remove;
 using Spectre.Console.Cli;
