+++ src/Tool/Git2SemVer.Tool/Commands/Changelog/ChangelogCommandSettings.cs
@@ -20,9 +20,19 @@ public class ChangelogCommandSettings : CommonCommandSettings
     [Description("Directory in which to place the generators metadata file.")]
     public string DataDirectory { get; set; } = "";
 
+    [CommandOption("-f|--force", IsHidden = true)]
+    [DefaultValue(false)]
+    [Description("Force create or update regardless if there a no new commits. For debugging.")]
+    public bool Force { get; set; }
+
     [CommandOption("--host-type <TYPE>")]
     [Description("Force the host type. Use for testing expected behaviour on other hosts. Valid values are 'Custom', 'Uncontrolled', 'TeamCity', or 'GitHub'.")]
     public string? HostType { get; set; } = null;
+
+    [CommandOption("-i|--incremental")]
+    [DefaultValue(true)]
+    [Description("Enable incremental changelog updating. Setting this to false will cause a new changelog to be created.")]
+    public bool Incremental { get; set; }
 
     [CommandOption("-o|--output <FILEPATH>")]
     [DefaultValue("CHANGELOG.md")]
@@ -38,14 +48,4 @@ public class ChangelogCommandSettings : CommonCommandSettings
     [DefaultValue(false)]
     [Description("Enable writing generated changelog to the console.")]
     public bool WriteToConsole { get; set; }
-
-    [CommandOption("-i|--incremental")]
-    [DefaultValue(true)]
-    [Description("Enable incremental changelog updating. Setting this to false will cause the changelog to be overwritten.")]
-    public bool Incremental { get; set; }
-
-    [CommandOption("-f|--force", IsHidden = true)]
-    [DefaultValue(false)]
-    [Description("Force create or update regardless if there a no new commits. For debugging.")]
-    public bool Force { get; set; }
 }
