+++ src/Tool/Git2SemVer.Tool/Commands/CommandFactory.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using NoeticTools.Git2SemVer.Tool.Commands.Add;
 using NoeticTools.Git2SemVer.Tool.Commands.Changelog;
-using NoeticTools.Git2SemVer.Tool.Commands.Remove;
-using NoeticTools.Git2SemVer.Tool.Commands.Run;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Add;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Remove;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Run;
 
 
 // ReSharper disable ClassNeverInstantiated.Global
