+++ src/Tool/Git2SemVer.Tool/Commands/ICommandFactory.cs
@@ -1,7 +1,7 @@
-﻿using NoeticTools.Git2SemVer.Tool.Commands.Add;
-using NoeticTools.Git2SemVer.Tool.Commands.Changelog;
-using NoeticTools.Git2SemVer.Tool.Commands.Remove;
-using NoeticTools.Git2SemVer.Tool.Commands.Run;
+﻿using NoeticTools.Git2SemVer.Tool.Commands.Changelog;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Add;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Remove;
+using NoeticTools.Git2SemVer.Tool.Commands.Versioning.Run;
 
 
 namespace NoeticTools.Git2SemVer.Tool.Commands;
