+++ src/Tool/Git2SemVer.Tool/Commands/Versioning/Add/AddCliCommand.cs
@@ -6,9 +6,9 @@ using Spectre.Console.Cli;
 
 // ReSharper disable ClassNeverInstantiated.Global
 
-namespace NoeticTools.Git2SemVer.Tool.Commands.Remove;
+namespace NoeticTools.Git2SemVer.Tool.Commands.Versioning.Add;
 
-internal class RemoveCliCommand : Command<SolutionCommandSettings>
+internal class AddCliCommand : Command<SolutionCommandSettings>
 {
     public override int Execute(CommandContext context, SolutionCommandSettings settings)
     {
@@ -17,7 +17,7 @@ internal class RemoveCliCommand : Command<SolutionCommandSettings>
         var commandFactory = serviceProvider.GetService<ICommandFactory>()!;
 
         console.Unattended = settings.Unattended;
-        var runner = commandFactory.CreateRemoveCommand();
+        var runner = commandFactory.CreateAddCommand();
         runner.Execute(settings.SolutionName, settings.Unattended);
         return (int)(runner.HasError ? ReturnCodes.CommandError : ReturnCodes.Succeeded);
     }
