+++ src/Tool/Git2SemVer.Tool/Commands/Versioning/Run/IRunCommand.cs
@@ -1,4 +1,4 @@
-﻿namespace NoeticTools.Git2SemVer.Tool.Commands.Run;
+﻿namespace NoeticTools.Git2SemVer.Tool.Commands.Versioning.Run;
 
 internal interface IRunCommand
 {
